Move polymer.json to root folder
This makes sure commands like `polymer lint` works again
