+++ gulp/project.js
@@ -21,7 +21,7 @@ const cssSlam = require('css-slam').gulp;
 const htmlMinifier = require('gulp-html-minifier');
 
 const swPrecacheConfig = require('./sw-precache-config.js');
-const polymerJson = require('./polymer.json');
+const polymerJson = require('../polymer.json');
 const polymerProject = new polymerBuild.PolymerProject(polymerJson);
 const buildDirectory = './build';
